renamed EditorSettings component to PostOptions
